+++ govcd/api_vcd_test.go
@@ -850,7 +850,7 @@ func (vcd *TestVCD) removeLeftoverEntities(entity CleanupEntity) {
 
 		vcd.infoCleanup(removedMsg, entity.EntityType, entity.Name, entity.CreatedBy)
 	case "OpenApiEntityGlobalDefaultSegmentProfileTemplate":
-		_, err := vcd.client.UpdateGlobalDefaultSegmentProfileTemplates(&types.NsxtSegmentProfileTemplateDefaultDefinition{})
+		_, err := vcd.client.UpdateGlobalDefaultSegmentProfileTemplates(&types.NsxtGlobalDefaultSegmentProfileTemplate{})
 		if err != nil {
 			vcd.infoCleanup(notDeletedMsg, entity.EntityType, entity.Name, err)
 			return
